requirements.txt
pandas
openpyxl
python-docx
pdfplumber
rapidfuzz
numpy
langgraph
langchain
langchain-google-genai
pydantic
python-dotenv
langsmith
